Add tests for quality assurance
